requirements.txt
lvis==0.5.3
mmcv-full==1.7.1
mmdet==2.28.1
mmengine==0.8.5
opencv-python==4.8.0.76
Pillow==10.0.0
pycocotools==2.0.7
timm==0.9.7
tokenizers==0.13.3
transformers==4.33.0
triton==2.0.0
webdataset>=0.2.5
yapf==0.40.1
xformers==0.0.18
einops
regex
apex
ftfy
tqdm
pandas
braceexpand
huggingface_hub
fsspec
openpyxl